refactor(api): strangle main settings into a use-case (ADR-0020, WS-A2)
handlers_settings.go carried ~37 inline s.config reaches: getSettings assembled a
20-key read model from config under RLock, and updateSettings did the optimistic
ETag check + 6 applyXxx mutations + save, with ~600 LOC of apply helpers in the
transport layer. Move it behind a clean-hexagonal application service.

- New internal/settings/management: Service over a Store port (Read runs fn under
  the config RLock; Write runs fn under the write lock then saves on nil error —
  the #783 unlock-before-save pattern). Get() builds the read model + ETag;
  Update(updates, ifMatch) does the atomic ETag compare-and-apply, returning typed
  ErrConflict (412) / ErrValidation (400). All applyXxx helpers + buildThreshold/
  buildCardSettings moved verbatim (behavior-identical); domain-meaningful naming
  per ADR-0020 §2 (Service/Store, never *UseCase). Extracted tier-key constants to
  satisfy goconst on the moved code.
- Adapter in internal/app/settings.go: NewSettingsManagement + managementStore over
  the live config + path.
- handlers_settings.go: getSettings/updateSettings now thin (decode → service →
  encode), mapping the typed errors to HTTP and keeping the SaveToActiveProfile +
  fresh-ETag emit. 1072→258 LOC. Deleted the 6 applyXxx + ~15 threshold helpers.
- testing.go harness wires settingsManagement (mirrors settingsStore) so the
  api tests construct it (caught a nil-deref in TestHandleSettingsGET).

Tests: management service (Get map+etag, Update apply, ErrConflict, ErrValidation).
Behavior-preserving. Non-race api suite + targeted -race on the changed surface +
filename/route/escaping/lint gates all green. (Full-package -race timed out on a
known Go 1.26.4 race-GC environmental flake under load; changed surface is
race-clean.)
