+++ internal/app/settings.go
@@ -1,10 +1,11 @@
 package app
 
-// settings.go wires the composition root to the settings-persistence
-// application (use-case) service (ADR-0020). The adapters below implement the
-// narrow ports declared in internal/settings/persistence over the concrete
-// database repositories and the live config, so the settings handlers depend on
-// a use-case instead of reaching into the server's service fields / s.db() directly.
+// settings.go wires the composition root to the settings-persistence and
+// settings-management application (use-case) services (ADR-0020). The adapters
+// below implement the narrow ports declared in internal/settings/persistence
+// and internal/settings/management over the concrete collaborators (config,
+// database), so the API handlers depend on use-cases instead of reaching into
+// the server's service fields directly.
 
 import (
 	"context"
@@ -12,6 +13,7 @@ import (
 
 	"github.com/MustardSeedNetworks/seed/internal/config"
 	"github.com/MustardSeedNetworks/seed/internal/database"
+	"github.com/MustardSeedNetworks/seed/internal/settings/management"
 	"github.com/MustardSeedNetworks/seed/internal/settings/persistence"
 )
 
@@ -75,4 +77,38 @@ type settingsConfigSource struct {
 
 func (c settingsConfigSource) ProfileJSON() (string, error) {
 	return c.cfg.ToProfileJSON()
+}
+
+// NewSettingsManagement builds the settings-management use-case (ADR-0020,
+// WS-A2) from the live config and the on-disk config path. cfg and path are
+// fixed for the process lifetime.
+func NewSettingsManagement(cfg *config.Config, path string) *management.Service {
+	return management.NewService(managementStore{cfg: cfg, path: path})
+}
+
+// managementStore implements management.Store over the live config, owning
+// the lock + on-disk save the port abstracts away.
+type managementStore struct {
+	cfg  *config.Config
+	path string
+}
+
+// Read calls fn with the live config held under the config RLock.
+func (s managementStore) Read(fn func(*config.Config)) {
+	s.cfg.RLock()
+	defer s.cfg.RUnlock()
+	fn(s.cfg)
+}
+
+// Write calls fn with the live config held under the config Lock. The lock is
+// released before Save acquires its own RLock to avoid the historic deadlock
+// (fixes #783). If fn returns a non-nil error the config is not saved.
+func (s managementStore) Write(fn func(*config.Config) error) error {
+	s.cfg.Lock()
+	err := fn(s.cfg)
+	s.cfg.Unlock()
+	if err != nil {
+		return err
+	}
+	return s.cfg.Save(s.path)
 }
